PDU queue out PDUs in order of itt only if the outqueue is at risk to block
Previous commit 1f4a66a sorted all PDUs in order of their itt. In fact
the only command that could be blocked would have been a Data-Out PDU
(or a retransmitted PDU). Sorting all PDUs according to itt lead to
all Data-OUT PDUs being queued before consecutive Command PDUs.
This turned out to be a performance drawback.

We switched the behaviour to only queue PDUs in front places of the queue
if there is an PDU in the queue that would block the outqueue (leading
to a deadlock).

Signed-off-by: Peter Lieven <pl@kamp.de>